requirements.txt
# Only needed if you enable real GPT calls
python-dotenv>=1.0.1
openai>=1.0.0
